+++ sql/mysql/Oracle/CSharp/MySQLParserBase.cs
@@ -18,7 +18,7 @@ public abstract class MySQLParserBase : Parser {
     protected MySQLParserBase(ITokenStream input, TextWriter output, TextWriter errorOutput) : base(input, output, errorOutput)
     {
         this.serverVersion = 80200;
-        this.sqlModeFromString("ANSI_QUOTES");
+        this.sqlModes = SqlModes.sqlModeFromString("ANSI_QUOTES");
     }
 
     /**
@@ -50,36 +50,5 @@ public abstract class MySQLParserBase : Parser {
     public bool isSelectStatementWithInto()
     {
         return serverVersion >= 80024 && serverVersion < 80031;
-    }
-
-    /**
-     * Converts a mode string into individual mode flags.
-     *
-     * @param modes The input string to parse.
-     */
-    public void sqlModeFromString(string modes)
-    {
-        this.sqlModes = new HashSet<SqlMode>();
-
-        var parts = modes.ToUpper().Split(",");
-        foreach (var mode in parts)
-        {
-            if (mode == "ANSI" || mode == "DB2" || mode == "MAXDB" || mode == "MSSQL" || mode == "ORACLE" ||
-            mode == "POSTGRESQL") {
-                this.sqlModes.Add(SqlMode.AnsiQuotes);
-                this.sqlModes.Add(SqlMode.PipesAsConcat);
-                this.sqlModes.Add(SqlMode.IgnoreSpace);
-            } else if (mode == "ANSI_QUOTES") {
-                this.sqlModes.Add(SqlMode.AnsiQuotes);
-            } else if (mode == "PIPES_AS_CONCAT") {
-                this.sqlModes.Add(SqlMode.PipesAsConcat);
-            } else if (mode == "NO_BACKSLASH_ESCAPES") {
-                this.sqlModes.Add(SqlMode.NoBackslashEscapes);
-            } else if (mode == "IGNORE_SPACE") {
-                this.sqlModes.Add(SqlMode.IgnoreSpace);
-            } else if (mode == "HIGH_NOT_PRECEDENCE" || mode == "MYSQL323" || mode == "MYSQL40") {
-                this.sqlModes.Add(SqlMode.HighNotPrecedence);
-            }
-        }
     }
 }
