+++ sql/mysql/Oracle/CSharp/SqlMode.cs
@@ -1,0 +1,9 @@
+/** SQL modes that control parsing behavior. */
+public enum SqlMode {
+    NoMode,
+    AnsiQuotes,
+    HighNotPrecedence,
+    PipesAsConcat,
+    IgnoreSpace,
+    NoBackslashEscapes,
+}
